bo: Some modifications

+++ src/main/java/fr/eni/PizzaOnline/dal/CheeseDAO.java
@@ -2,6 +2,8 @@ package fr.eni.PizzaOnline.dal;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CheeseDAO extends CrudRepository<Cheese, Integer>{
+import fr.eni.PizzaOnline.bo.Cheese;
+
+public interface CheeseDAO extends CrudRepository<Cheese, Long>{
 
 }
